Split debug and perf versions of search table
