+++ hstrat/phylogenetic_inference/tree/__init__.py
@@ -3,6 +3,7 @@
 from . import trie_postprocess
 from ._build_tree import build_tree
 from ._build_tree_nj import build_tree_nj
+from ._build_tree_searchtable_debug import build_tree_searchtable_debug
 from ._build_tree_searchtable import build_tree_searchtable
 from ._build_tree_trie import build_tree_trie
 from ._build_tree_trie_ensemble import build_tree_trie_ensemble
@@ -22,6 +23,7 @@ __all__ = [
     "build_tree_upgma",
     "build_tree",
     "build_tree_searchtable",
+    "build_tree_searchtable_debug",
     "trie_postprocess",
     "build_trie_from_artifacts_matrix",
     "build_trie_from_artifacts_progressive",
@@ -35,6 +37,7 @@ _launder(
     [
         build_tree_nj,
         build_tree_searchtable,
+        build_tree_searchtable_debug,
         build_tree_trie,
         build_tree_trie_ensemble,
         build_tree_upgma,
